Move image loading to its own file
